backend

+++ src/main/java/com/example/EcoMonitoring/repository/CityRepository.java
@@ -6,12 +6,11 @@ import org.springframework.stereotype.Repository;
 
 import com.example.EcoMonitoring.model.City;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
 @Repository
-public interface CityRepository extends JpaRepository<City, Integer> {
+public interface CityRepository extends JpaRepository<City, Long> {
     Optional<City> findById(Long id);
     City findByName(String name);
 }
